+++ druid-shell/src/backend/x11/application.rs
@@ -36,8 +36,8 @@ use x11rb::xcb_ffi::XCBConnection;
 use crate::application::AppHandler;
 
 use super::clipboard::Clipboard;
-use super::util;
 use super::window::Window;
+use super::{util, xkb};
 
 // This creates a `struct WindowAtoms` containing the specified atoms as members (along with some
 // convenience methods to intern and query those atoms). We use the following atoms:
@@ -175,6 +175,7 @@ pub(crate) struct Application {
     render_argb32_pictformat_cursor: Option<Pictformat>,
     /// Newest timestamp that we received
     timestamp: Rc<Cell<Timestamp>>,
+    xkb_context: xkb::Context,
 }
 
 /// The mutable `Application` state.
@@ -183,6 +184,7 @@ struct State {
     quitting: bool,
     /// A collection of all the `Application` windows.
     windows: HashMap<u32, Rc<Window>>,
+    xkb_state: xkb::State,
 }
 
 #[derive(Clone, Debug)]
@@ -208,11 +210,27 @@ impl Application {
         // https://github.com/linebender/druid/pull/1025#discussion_r442777892
         let (conn, screen_num) = XCBConnection::connect(None)?;
         let rdb = Rc::new(ResourceDb::new_from_default(&conn)?);
+        let xkb_context = xkb::Context::new();
+        xkb_context.set_log_level(tracing::Level::DEBUG);
+        use x11rb::protocol::xkb::ConnectionExt;
+        conn.xkb_use_extension(1, 0)?
+            .reply()
+            .context("init xkb extension")?;
+        let device_id = xkb_context
+            .core_keyboard_device_id(&conn)
+            .context("get core keyboard device id")?;
+
+        let keymap = xkb_context
+            .keymap_from_device(&conn, device_id)
+            .context("key map from device")?;
+
+        let xkb_state = keymap.state();
         let connection = Rc::new(conn);
         let window_id = Application::create_event_window(&connection, screen_num)?;
         let state = Rc::new(RefCell::new(State {
             quitting: false,
             windows: HashMap::new(),
+            xkb_state,
         }));
 
         let (idle_read, idle_write) = nix::unistd::pipe2(nix::fcntl::OFlag::O_NONBLOCK)?;
@@ -331,6 +349,7 @@ impl Application {
             marker: std::marker::PhantomData,
             render_argb32_pictformat_cursor,
             timestamp,
+            xkb_context,
         })
     }
 
@@ -523,7 +542,25 @@ impl Application {
                 let w = self
                     .window(ev.event)
                     .context("KEY_PRESS - failed to get window")?;
-                w.handle_key_press(ev);
+                let hw_keycode = ev.detail;
+                let mut state = borrow_mut!(self.state)?;
+                let key_event = state
+                    .xkb_state
+                    .key_event(hw_keycode as _, keyboard_types::KeyState::Down);
+
+                w.handle_key_event(key_event);
+            }
+            Event::KeyRelease(ev) => {
+                let w = self
+                    .window(ev.event)
+                    .context("KEY_PRESS - failed to get window")?;
+                let hw_keycode = ev.detail;
+                let mut state = borrow_mut!(self.state)?;
+                let key_event = state
+                    .xkb_state
+                    .key_event(hw_keycode as _, keyboard_types::KeyState::Up);
+
+                w.handle_key_event(key_event);
             }
             Event::ButtonPress(ev) => {
                 let w = self